requirements.txt
discord.py
requests
python_decouple
